+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 (unreleased)
+ - Adding api package (@LeoCBS)
  - Adding storage/mongo.go package (@LeoCBS)